requirements.txt
python-jobspy
pandas
openpyxl
google-generativeai
googlesearch-python
python-dotenv
requests
